% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_engine.R
\name{grid_rho_search}
\alias{grid_rho_search}
\title{Grid search for bi-square objective minimum (starting value)}
\usage{
grid_rho_search(y, var_fn, k, width = 0.01)
}
\description{
Grid search for bi-square objective minimum (starting value)
}
\keyword{internal}
