% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_engine.R
\name{vcov_scaling_fn}
\alias{vcov_scaling_fn}
\title{Delta-method covariance matrix of the scaling functions}
\usage{
vcov_scaling_fn(mle, theta = NULL, scale_by = "pooled")
}
\arguments{
\item{mle}{A validated mle list.}

\item{theta}{Optional scalar evaluated under H0.}

\item{scale_by}{Passed to gradient function.}
}
\value{
A square matrix of size
  \code{n_items * n_thresholds}.
}
\description{
Delta-method covariance matrix of the scaling functions
}
\keyword{internal}
